requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
ultralytics>=8.1.0
Pillow>=10.0.0
httpx>=0.27.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
